+++ neo/lib/protocol.py
@@ -225,6 +225,7 @@ class Packet(object):
     _code = None
     _fmt = None
     _id = None
+    nodelay = True
     poll_thread = False
 
     def __init__(self, *args):
@@ -1444,6 +1445,8 @@ class AddTransaction(Packet):
     """
     S -> S
     """
+    nodelay = False
+
     _fmt = PStruct('add_transaction',
         PTID('tid'),
         PString('user'),
@@ -1483,6 +1486,8 @@ class AddObject(Packet):
     """
     S -> S
     """
+    nodelay = False
+
     _fmt = PStruct('add_object',
         POID('oid'),
         PTID('serial'),