+++ neo/scripts/neolog.py
@@ -22,7 +22,7 @@ from bisect import insort
 from logging import getLevelName
 from zlib import decompress
 
-comp_dict = dict(bz2=bz2.BZ2File, gz=gzip.GzipFile)
+comp_dict = dict(bz2=bz2.BZ2File, gz=gzip.GzipFile, xz='xzcat')
 
 class Log(object):
 
@@ -41,14 +41,17 @@ class Log(object):
             name, ext = name.rsplit(os.extsep, 1)
             ZipFile = comp_dict[ext]
         except (KeyError, ValueError):
-            # WKRD: Python does not support URI so we can't open in read-only
-            #       mode. See http://bugs.python.org/issue13773
+            # BBB: Python 2 does not support URI so we can't open in read-only
+            #      mode. See https://bugs.python.org/issue13773
             os.stat(db_path) # do not create empty DB if file is missing
             self._db = sqlite3.connect(db_path)
         else:
-            import shutil, tempfile
+            import shutil, subprocess, tempfile
             with tempfile.NamedTemporaryFile() as f:
-                shutil.copyfileobj(ZipFile(db_path), f)
+                if type(ZipFile) is str:
+                    subprocess.check_call((ZipFile, db_path), stdout=f)
+                else:
+                    shutil.copyfileobj(ZipFile(db_path), f)
                 self._db = sqlite3.connect(f.name)
             name = name.rsplit(os.extsep, 1)[0]
         self._default_name = name
@@ -241,17 +244,29 @@ def main():
     parser.add_option('-s', '--sleep-interval', type="float", default=1,
         help='with -f, sleep for approximately N seconds (default 1.0)'
               ' between iterations', metavar='N')
-    parser.add_option('--from', dest='filter_from', type="float",
+    parser.add_option('--from', dest='filter_from',
         help='show records more recent that timestamp N if N > 0,'
-             ' or now+N if N < 0', metavar='N')
+             ' or now+N if N < 0; N can also be a string that is'
+             ' parseable by dateutil ', metavar='N')
     options, args = parser.parse_args()
     if options.sleep_interval <= 0:
         parser.error("sleep_interval must be positive")
     if not args:
         parser.error("no log specified")
     filter_from = options.filter_from
-    if filter_from and filter_from < 0:
-        filter_from += time.time()
+    if filter_from:
+        try:
+            filter_from = float(options.filter_from)
+        except ValueError:
+            from dateutil.parser import parse
+            x = parse(filter_from)
+            if x.tzinfo:
+                filter_from = (x - x.fromtimestamp(0, x.tzinfo)).total_seconds()
+            else:
+                filter_from = time.mktime(x.timetuple()) + x.microsecond * 1e-6
+        else:
+            if filter_from < 0:
+                filter_from += time.time()
     node_list = options.node or []
     try:
         node_list.remove('-')