+++ neo/storage/database/mysqldb.py
@@ -172,7 +172,7 @@ class MySQLDatabaseManager(DatabaseManager):
             if e.args[0] != NO_SUCH_TABLE:
                 raise
 
-    def _setup(self):
+    def _setup(self, dedup=False):
         self._config.clear()
         q = self.query
         p = engine = self._engine
@@ -240,9 +240,9 @@ class MySQLDatabaseManager(DatabaseManager):
                  id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
                  hash BINARY(20) NOT NULL,
                  compression TINYINT UNSIGNED NULL,
-                 value MEDIUMBLOB NOT NULL,
-                 UNIQUE (hash, compression)
-             ) ENGINE=""" + engine)
+                 value MEDIUMBLOB NOT NULL%s
+             ) ENGINE=%s""" % (""",
+                 UNIQUE (hash, compression)""" if dedup else "", engine))
 
         q("""CREATE TABLE IF NOT EXISTS bigdata (
                  id INT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY,
@@ -701,6 +701,21 @@ class MySQLDatabaseManager(DatabaseManager):
              self._getPackTID(), offset, length))
         if r:
             return [(p64(tid), length or 0) for tid, length in r]
+
+    def _fetchObject(self, oid, tid):
+        r = self.query(
+            'SELECT tid, compression, data.hash, value, value_tid'
+            ' FROM obj FORCE INDEX(`partition`)'
+            ' LEFT JOIN data ON (obj.data_id = data.id)'
+            ' WHERE `partition` = %d AND oid = %d AND tid = %d'
+            % (self._getReadablePartition(oid), oid, tid))
+        if r:
+            r = r[0]
+            compression = r[1]
+            if compression and compression & 0x80:
+                return (r[0], compression & 0x7f, r[2],
+                    ''.join(self._bigData(data)), r[4])
+            return r
 
     def getReplicationObjectList(self, min_tid, max_tid, length, partition,
             min_oid):